FIREFLY-1895: Search title improvements
