+++ src/firefly/js/tables/TableUtil.js
@@ -1631,6 +1631,26 @@ export function ensureEnumVals(tableModel) {
     });
 }
 
+/**
+ * smartly decides the next tbl title given the existing table titles, to avoid duplicates
+ * @param {string} [inTitle] the default title
+ * @return {string} returns a numbered title (if the given title already exists), else return the default title as is
+ */
+export function makeNumberedTitle(inTitle) {
+    if (getTblIdsByGroup().map((tbl_id) => getTblById(tbl_id)?.title).every( (t) => t!==inTitle)) {
+        return inTitle;
+    }
+    const numList = getTblIdsByGroup()
+        .map((tbl_id) => getTblById(tbl_id)?.title)
+        .filter((title) => title && title.startsWith(inTitle))
+        .map((title) => title?.substring(inTitle.length).trim().split('-')?.[1])
+        .map(Number)
+        .filter(Boolean);
+
+    const maxNum = numList?.length ? Math.max(...numList) : 0;
+    return inTitle + ` - ${maxNum + 1}`;
+}
+
 
 /*-------------------------------------private------------------------------------------------*/
